folder restructure for client and server
