+++ client/.eslintcache
@@ -1,0 +1,1 @@
+[{"C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\index.js":"1","C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\App.js":"2","C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\Expenses\\Expenses.js":"3","C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\NewExpense\\NewExpense.js":"4","C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\Expenses\\ExpensesList.js":"5","C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\Expenses\\ExpensesFilter.js":"6","C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\Expenses\\ExpensesChart.js":"7","C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\NewExpense\\ExpenseForm.js":"8","C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\Ui\\Card.js":"9","C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\Expenses\\ExpenseItem.js":"10","C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\Chart\\Chart.js":"11","C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\Expenses\\ExpenseDate.js":"12","C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\Chart\\ChartBar.js":"13"},{"size":148,"mtime":1650127777025,"results":"14","hashOfConfig":"15"},{"size":1124,"mtime":1650127777010,"results":"16","hashOfConfig":"15"},{"size":919,"mtime":1650128914677,"results":"17","hashOfConfig":"15"},{"size":950,"mtime":1650127777022,"results":"18","hashOfConfig":"15"},{"size":570,"mtime":1650127777020,"results":"19","hashOfConfig":"15"},{"size":689,"mtime":1650127777019,"results":"20","hashOfConfig":"15"},{"size":819,"mtime":1650127777018,"results":"21","hashOfConfig":"15"},{"size":2606,"mtime":1650127777021,"results":"22","hashOfConfig":"15"},{"size":216,"mtime":1650127777024,"results":"23","hashOfConfig":"15"},{"size":513,"mtime":1650128523187,"results":"24","hashOfConfig":"15"},{"size":568,"mtime":1650127777012,"results":"25","hashOfConfig":"15"},{"size":551,"mtime":1650127777014,"results":"26","hashOfConfig":"15"},{"size":560,"mtime":1650127777012,"results":"27","hashOfConfig":"15"},{"filePath":"28","messages":"29","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"m42tl4",{"filePath":"30","messages":"31","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"42","messages":"43","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"44","messages":"45","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"46","messages":"47","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"48","messages":"49","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"50","messages":"51","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"52","messages":"53","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\index.js",[],"C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\App.js",[],"C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\Expenses\\Expenses.js",[],"C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\NewExpense\\NewExpense.js",[],"C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\Expenses\\ExpensesList.js",[],"C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\Expenses\\ExpensesFilter.js",[],"C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\Expenses\\ExpensesChart.js",[],"C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\NewExpense\\ExpenseForm.js",[],"C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\Ui\\Card.js",[],"C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\Expenses\\ExpenseItem.js",[],"C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\Chart\\Chart.js",[],"C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\Expenses\\ExpenseDate.js",[],"C:\\Users\\kcd25\\OneDrive\\Documents\\Projects\\finance-tracker\\client\\src\\components\\Chart\\ChartBar.js",[]]
