+++ client/src/components/Expenses/ExpenseItem.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import ExpenseDate from './ExpenseDate';
-import Card from '../UI/Card';
+import Card from '../Ui/Card';
 import './ExpenseItem.css';
 
 const ExpenseItem = (props) => {
